Base for a new Module-based compilation process
Right now our JavaScript code is executed on the global scope, with a
prologue and epilogue that we insert to mimic a module system. Including
another file is done through the require() function.

JavaScript has since obtained a powerful module system, to which we
should switch over. This requires a very different manner of loading the
code. An experimental Modulator is therefore being developed.